requirements.txt
aiohttp
fluxer.py
py-cord
python-telegram-bot
requests